requirements.txt
django>=4.2
djangorestframework>=3.14
psycopg2-binary>=2.9
gunicorn>=21.0
twilio
python-dotenv
djangorestframework-simplejwt>=5.5.1
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.1
django-redis>=5.4.0
drf_yasg
django-filter